Polymorphism! Yay
Made a Cookable interface, that is implemented on Food and Nutrient. The
mandatory Overrides are done in Mineral and Organic. The interesting
overrides are one step further down in Protein, Carbohydrate, Calcium,
and Iron. The first two work, Iron has no override on purpose. Calcium
reveals an error with another unrelated code portion.